Update CRE SDK and its capabilities with the renamed sdk -> cre package
