+++ core/scripts/cre/environment/examples/workflows/v2/cron/go.mod
@@ -3,8 +3,8 @@ module main
 go 1.24.4
 
 require (
-	github.com/smartcontractkit/cre-sdk-go v0.1.0
-	github.com/smartcontractkit/cre-sdk-go/capabilities/scheduler/cron v0.1.0
+	github.com/smartcontractkit/cre-sdk-go v0.2.1-0.20250729191525-ac1867f3ff34
+	github.com/smartcontractkit/cre-sdk-go/capabilities/scheduler/cron v0.2.1-0.20250729191525-ac1867f3ff34
 )
 
 require (
@@ -14,8 +14,8 @@ require (
 	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
 	github.com/rogpeppe/go-internal v1.12.0 // indirect
 	github.com/shopspring/decimal v1.4.0 // indirect
-	github.com/smartcontractkit/chainlink-common/pkg/values v0.0.0-20250707133321-27faefc9ce45 // indirect
-	github.com/smartcontractkit/chainlink-common/pkg/workflows/sdk/v2/pb v0.0.0-20250707133321-27faefc9ce45 // indirect
+	github.com/smartcontractkit/chainlink-common/pkg/values v0.0.0-20250718143957-41236f9ef8b4 // indirect
+	github.com/smartcontractkit/chainlink-common/pkg/workflows/sdk/v2/pb v0.0.0-20250718162715-0c70aa0d3a06 // indirect
 	github.com/stretchr/testify v1.10.0 // indirect
 	google.golang.org/protobuf v1.36.6 // indirect
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
