+++ core/services/workflows/cmd/cre/examples/v2/http_read/main.go
@@ -5,26 +5,26 @@ package main
 import (
 	"github.com/smartcontractkit/cre-sdk-go/capabilities/networking/http"
 	"github.com/smartcontractkit/cre-sdk-go/capabilities/scheduler/cron"
-	"github.com/smartcontractkit/cre-sdk-go/sdk"
-	"github.com/smartcontractkit/cre-sdk-go/sdk/wasm"
+	"github.com/smartcontractkit/cre-sdk-go/cre"
+	"github.com/smartcontractkit/cre-sdk-go/cre/wasm"
 )
 
-func RunSimpleCronWorkflow(_ *sdk.Environment[struct{}]) (sdk.Workflow[struct{}], error) {
+func RunSimpleCronWorkflow(_ *cre.Environment[struct{}]) (cre.Workflow[struct{}], error) {
 	cfg := &cron.Config{
 		Schedule: "*/3 * * * * *", // every 3 seconds
 	}
 
-	return sdk.Workflow[struct{}]{
-		sdk.Handler(
+	return cre.Workflow[struct{}]{
+		cre.Handler(
 			cron.Trigger(cfg),
 			onTrigger,
 		),
 	}, nil
 }
 
-func onTrigger(env *sdk.Environment[struct{}], runtime sdk.Runtime, outputs *cron.Payload) (string, error) {
+func onTrigger(env *cre.Environment[struct{}], runtime cre.Runtime, outputs *cron.Payload) (string, error) {
 	env.Logger.Info("onTrigger called")
-	ret, err := sdk.RunInNodeMode(env, runtime, func(env *sdk.NodeEnvironment[struct{}], nrt sdk.NodeRuntime) (string, error) {
+	ret, err := cre.RunInNodeMode(env, runtime, func(env *cre.NodeEnvironment[struct{}], nrt cre.NodeRuntime) (string, error) {
 		httpClient := http.Client{}
 		resp, err := httpClient.SendRequest(nrt, &http.Request{
 			Method:  "GET",
@@ -32,7 +32,7 @@ func onTrigger(env *sdk.Environment[struct{}], runtime sdk.Runtime, outputs *cro
 			Headers: map[string]string{"Content-Type": "application/json"},
 		}).Await()
 		return string(resp.Body), err
-	}, sdk.ConsensusIdenticalAggregation[string]()).Await()
+	}, cre.ConsensusIdenticalAggregation[string]()).Await()
 
 	if err != nil {
 		env.Logger.Error("Error in RunInNodeMode", "err", err)
