+++ core/services/workflows/cmd/cre/examples/v2/simple_cron/main.go
@@ -4,24 +4,24 @@ package main
 
 import (
 	"github.com/smartcontractkit/cre-sdk-go/capabilities/scheduler/cron"
-	"github.com/smartcontractkit/cre-sdk-go/sdk"
-	"github.com/smartcontractkit/cre-sdk-go/sdk/wasm"
+	"github.com/smartcontractkit/cre-sdk-go/cre"
+	"github.com/smartcontractkit/cre-sdk-go/cre/wasm"
 )
 
-func RunSimpleCronWorkflow(_ *sdk.Environment[struct{}]) (sdk.Workflow[struct{}], error) {
+func RunSimpleCronWorkflow(_ *cre.Environment[struct{}]) (cre.Workflow[struct{}], error) {
 	cfg := &cron.Config{
 		Schedule: "*/3 * * * * *", // every 3 seconds
 	}
 
-	return sdk.Workflow[struct{}]{
-		sdk.Handler(
+	return cre.Workflow[struct{}]{
+		cre.Handler(
 			cron.Trigger(cfg),
 			onTrigger,
 		),
 	}, nil
 }
 
-func onTrigger(env *sdk.Environment[struct{}], runtime sdk.Runtime, outputs *cron.Payload) (string, error) {
+func onTrigger(env *cre.Environment[struct{}], runtime cre.Runtime, outputs *cron.Payload) (string, error) {
 	env.Logger.Info("inside onTrigger handler")
 	return "success!", nil
 }
