+++ core/services/workflows/cmd/cre/examples/v2/simple_cron_with_config/main.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 
 	"github.com/smartcontractkit/cre-sdk-go/capabilities/scheduler/cron"
-	"github.com/smartcontractkit/cre-sdk-go/sdk"
-	"github.com/smartcontractkit/cre-sdk-go/sdk/wasm"
+	"github.com/smartcontractkit/cre-sdk-go/cre"
+	"github.com/smartcontractkit/cre-sdk-go/cre/wasm"
 	"gopkg.in/yaml.v3"
 )
 
@@ -15,20 +15,20 @@ type runtimeConfig struct {
 	Schedule string `yaml:"schedule"`
 }
 
-func RunSimpleCronWorkflow(env *sdk.Environment[*runtimeConfig]) (sdk.Workflow[*runtimeConfig], error) {
+func RunSimpleCronWorkflow(env *cre.Environment[*runtimeConfig]) (cre.Workflow[*runtimeConfig], error) {
 	cfg := &cron.Config{
 		Schedule: env.Config.Schedule,
 	}
 
-	return sdk.Workflow[*runtimeConfig]{
-		sdk.Handler(
+	return cre.Workflow[*runtimeConfig]{
+		cre.Handler(
 			cron.Trigger(cfg),
 			onTrigger,
 		),
 	}, nil
 }
 
-func onTrigger(env *sdk.Environment[*runtimeConfig], runtime sdk.Runtime, outputs *cron.Payload) (string, error) {
+func onTrigger(env *cre.Environment[*runtimeConfig], runtime cre.Runtime, outputs *cron.Payload) (string, error) {
 	env.Logger.Info("inside onTrigger handler")
 	return fmt.Sprintf("success (Schedule: %s)", env.Config.Schedule), nil
 }
